test something

+++ training/sampling.py
@@ -106,9 +106,9 @@ def save_test_samples(example_dir, data_loader, audio_mel_converter, models, los
         # secret_gen
         secret_z = torch.randn(mel.shape[0], noise_dim).to(device)
         # fake_secret_male = Variable(LongTensor(mel.size(0)).fill_(1.0), requires_grad=False).to(device)
-        fake_secret_male = torch.ones(mel.size(0), requires_grad=False).to(device)
+        fake_secret_male = torch.ones(mel.size(0), requires_grad=False, dtype=torch.int64).to(device)
         # fake_secret_female = Variable(LongTensor(mel.size(0)).fill_(0.0), requires_grad=False).to(device)
-        fake_secret_female = torch.zeros(mel.size(0), requires_grad=False).to(device)
+        fake_secret_female = torch.zeros(mel.size(0), requires_grad=False, dtype=torch.int64).to(device)
         fake_mel_male = models['secret_gen'](filtered, secret_z, fake_secret_male).detach()
         fake_mel_female = models['secret_gen'](filtered, secret_z, fake_secret_female).detach()
